Migrate boxing & default value renderers away from StringTemplate.
This is part of the KotlinPoet migration; really we're just moving the codegen to format strings in Kotlin since the generated code for each of these functions is small and the main complexity is in the conditionals. Also, KotlinPoet doesn't really support writing code blocks in a way that would be useful here.